+++ docs/changelog.md
@@ -28,6 +28,7 @@ Click on the version in the following list to see all changes since the previous
 * refactor: introduce the concept of a card to allow further logic
 * refactor: optimize imports
 * refactor: improve readability of mapper for REST representation of a session
+* refactor: rename estimate service in preparation for new custom field logic
 
 ### [4.10.0] - 2020-03-28